Create Energy
Energy drink calculations are simple - to decide how many drinks you should have, you are given a number. Find the sum of the even digits and the sum of the odd digits of that number, then compare these sums and:

If the sum of the even digits is bigger, drink energy drinks.
If the sum of the odd digits is bigger, drink cups of coffee.
If the two sums are equal, drink both.