Create Alone Numbers
We will give you an array and a target! You need to find all "alone" elements in the array that match the given target. These elements are alone if they have values before and after them, but those values are different from them.

Return a new version of the given array where every element that matches the target and is alone is replaced by whichever value to its left or right is larger.